Get a song lyric by Lupicínio every day

Draw a random song lyric by Brazilian musician Lupicínio Rodrigues from <www.letras.mus.br/> based on an seed.

Usage

Lupicinio(max.char = NULL, flag = " ", seed = NULL)

Arguments

max.char

integer; maximum number of characters to be returned; if NULL (DEFAULT) the text is not cropped.

flag

character; if max.char is used, everything after the last time this character appears will be removed; the DEFAULT is a single space " ".

seed

numeric; a seed for the draw; if NULL (DEFAULT) no seed is set.

Value

list; contains author, title, text, and source.

Examples

Lupicinio(max.char = 1000)
#> $author
#> [1] "Lupicínio Rodrigues"
#> 
#> $title
#> [1] "Margarida"
#> 
#> $text
#> [1] "Ai, Margarida\n\nMargarida meu amor\nSe os anjos do céu são loiros\nEla é um anjo sim senhor\n\nEu vou contar pra vocês\nO jeito da Margarida\nÉ uma espiuinha de milho\nNo ponto de ser colhida\n\nAi, Margarida\n\nA gente quando se espelha\nNos olhos da Margarida\nVê duas pombinhas brancas\nBeliscando nossas vidas\n\nAi, Margarida\n\nQuando ela toma banho\nNa praia toda encolhida\nAs ondas do mar se curvam\nAnte as curvas da Margarida\n\nAi, ..."
#> 
#> $source
#> [1] "https://www.letras.mus.br/lupcinio-rodrigues/1410842/"
#> 
# should return the first 1000 ish characters of today's Lupicínio song lyric.